+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project (now) adheres to [Calendar Versioning](https://calver.org/#sche
   - Custom target for OSC messages (e.g. for use with Resonite, OSC routers, or when OSCQuery is not available)
   - VRChat (OSCQuery) target (allows for disabling OSC messages being sent to VRChat)
 
+### Fixed
+
+- Disabled CEF disk cache
+
 ## [25.6.12]
 
 ### Added